refactor status file names
